# docs/api/openapi-tts.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "e-voice TTS",
    "description": "CPU-first streaming text-to-speech. Text in (whole or token by token), audio out while it is generated: every backend streams frame by frame. Voices are learned from a few seconds of audio (`/v1/voices`) and used by id in every endpoint.",
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/voices": {
      "get": {
        "tags": [
          "voices"
        ],
        "summary": "Learned voices, usable as `voice` in every endpoint.",
        "description": "# Errors\nThe voices directory cannot be read.",
        "operationId": "list",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceList"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "voices"
        ],
        "summary": "Learns a voice from one or more clips of the same speaker (multipart: `voice_id`, then `file`\nparts in any container `ffmpeg` would call common: wav, mp3, m4a, flac, ogg, webm). Clean speech,\n5-30 s in total; anything past 30 s is ignored. Replaces a voice with the same id.",
        "description": "# Errors\nA malformed form, undecodable audio, a backend that cannot learn voices, or a failed write.",
        "operationId": "create",
        "requestBody": {
          "description": "`voice_id` text field and one or more `file` parts",
          "content": {
            "multipart/form-data": {}
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceCreated"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "501": {
            "description": "the backend cannot learn voices",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/voices/{voice_id}": {
      "delete": {
        "tags": [
          "voices"
        ],
        "summary": "Forgets a learned voice.",
        "description": "# Errors\nAn invalid or unknown id.",
        "operationId": "remove",
        "parameters": [
          {
            "name": "voice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {},
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "description": "Error body in the OpenAI shape: `{\"error\": {\"message\", \"type\"}}`.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "message",
          "type"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "VoiceCreated": {
        "type": "object",
        "required": [
          "voice_id",
          "seconds"
        ],
        "properties": {
          "seconds": {
            "type": "number",
            "format": "double",
            "description": "Seconds of audio the voice was learned from."
          },
          "voice_id": {
            "type": "string"
          }
        }
      },
      "VoiceList": {
        "type": "object",
        "required": [
          "voices"
        ],
        "properties": {
          "voices": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  },
...
}